# gui/backend/openapi.yaml
openapi: 3.1.0
info:
  title: element-redis GUI API
  version: 0.1.0
servers:
  - url: /
paths:
  /api/v1/namespaces:
    get:
      operationId: namespaces
      responses:
        "200":
          description: Allowed Redis namespaces (prefix families) from preset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespacesResponse"
  /api/v1/config:
    get:
      operationId: config
      responses:
        "200":
          description: Runtime UI/API configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
  /api/v1/bitmaps:
    get:
      operationId: bitmaps
      parameters:
        - in: query
          name: ns
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Bit dictionary resolved from preset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BitmapsResponse"
    put:
      operationId: bitmapsPut
      parameters:
        - in: query
          name: ns
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BitmapsDocument"
      responses:
        "200":
          description: Saved bit-maps document and returned resolved view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BitmapsResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/health:
    get:
      operationId: health
      responses:
        "200":
          description: Health status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/v1/elements/put:
    post:
      operationId: elementsPut
      parameters:
        - in: query
          name: ns
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutRequest"
      responses:
        "200":
          description: Put element
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/elements/get:
    get:
      operationId: elementsGet
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - in: query
          name: ns
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 200
            minimum: 1
            maximum: 4096
      responses:
        "200":
          description: Get element bits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/query:
    post:
      operationId: query
      parameters:
        - in: query
          name: ns
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/store:
    post:
      operationId: store
      parameters:
        - in: query
          name: ns
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreRequest"
      responses:
        "200":
          description: Store results with TTL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: storeDelete
      parameters:
        - in: query
          name: store_key
          required: true
          schema:
            type: string
        - in: query
          name: ns
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Delete store key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteStoreResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/store/inspect:
    get:
      operationId: storeInspect
      parameters:
        - in: query
          name: store_key
          required: true
          schema:
            type: string
        - in: query
          name: ns
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 200
            minimum: 1
            maximum: 5000
      responses:
        "200":
          description: Inspect store key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectStoreResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/logs:
    get:
      operationId: logs
      parameters:
        - in: query
          name: tail
          required: false
          schema:
            type: integer
            default: 200
            minimum: 1
            maximum: 2000
      responses:
        "200":
          description: Tail backend logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"
components:
  schemas:
    Bit:
      type: integer
      minimum: 0
      maximum: 4095
    Name:
      type: string
      minLength: 1
      maxLength: 100
    ErrorObject:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          const: false
        error:
          $ref: "#/components/schemas/ErrorObject"
    PutRequest:
      type: object
      required: [name, bits]
      properties:
        ns:
          type: string
          nullable: true
        name:
          $ref: "#/components/schemas/Name"
        bits:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Bit"
    HealthResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [backend_version, redis]
          properties:
            backend_version:
              type: string
            preset:
              type: string
              nullable: true
            preset_path:
              type: string
              nullable: true
            redis:
              type: object
              required: [ok, ping_ms, used_memory]
              properties:
                ok:
                  type: boolean
                ping_ms:
                  type: integer
                used_memory:
                  type: integer
                  nullable: true
    PutResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [name, written_bits]
          properties:
            name:
              $ref: "#/components/schemas/Name"
            written_bits:
              type: integer
              minimum: 1
    GetResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [name, bits, count, returned, limit]
          properties:
            name:
              $ref: "#/components/schemas/Name"
            bits:
              type: array
              items:
                $ref: "#/components/schemas/Bit"
            count:
              type: integer
            returned:
              type: integer
            limit:
              type: integer
    QueryRequest:
      oneOf:
        - type: object
          required: [type, bit]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find
            bit:
              $ref: "#/components/schemas/Bit"
            limit:
              type: integer
              default: 200
              minimum: 1
              maximum: 5000
        - type: object
          required: [type, bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_all
            bits:
              type: array
              minItems: 2
              items:
                $ref: "#/components/schemas/Bit"
            limit:
              type: integer
              default: 200
              minimum: 1
              maximum: 5000
        - type: object
          required: [type, bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_any
            bits:
              type: array
              minItems: 2
              items:
                $ref: "#/components/schemas/Bit"
            limit:
              type: integer
              default: 200
              minimum: 1
              maximum: 5000
        - type: object
          required: [type, include_bit, exclude_bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_not
            include_bit:
              $ref: "#/components/schemas/Bit"
            exclude_bits:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/Bit"
            limit:
              type: integer
              default: 200
              minimum: 1
              maximum: 5000
        - type: object
          required: [type, exclude_bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_universe_not
            exclude_bits:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/Bit"
            limit:
              type: integer
              default: 200
              minimum: 1
              maximum: 5000
    QueryResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [ns, type, count, returned, limit, names]
          properties:
            ns:
              type: string
            type:
              type: string
            count:
              type: integer
            returned:
              type: integer
            limit:
              type: integer
            names:
              type: array
              items:
                type: string
    StoreRequest:
      oneOf:
        - type: object
          required: [type, ttl_sec, bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_all_store
            ttl_sec:
              type: integer
              exclusiveMinimum: 0
              maximum: 86400
            bits:
              type: array
              minItems: 2
              items:
                $ref: "#/components/schemas/Bit"
        - type: object
          required: [type, ttl_sec, bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_any_store
            ttl_sec:
              type: integer
              exclusiveMinimum: 0
              maximum: 86400
            bits:
              type: array
              minItems: 2
              items:
                $ref: "#/components/schemas/Bit"
        - type: object
          required: [type, ttl_sec, include_bit, exclude_bits]
          properties:
            ns:
              type: string
              nullable: true
            type:
              const: find_not_store
            ttl_sec:
              type: integer
              exclusiveMinimum: 0
              maximum: 86400
            include_bit:
              $ref: "#/components/schemas/Bit"
            exclude_bits:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/Bit"
    StoreResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [ns, store_key, ttl_remaining, count, preview_limit, preview]
          properties:
            ns:
              type: string
            store_key:
              type: string
            ttl_remaining:
              type: integer
            count:
              type: integer
            preview_limit:
              type: integer
            preview:
              type: array
              items:
                type: string
    InspectStoreResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [ns, store_key, ttl_remaining, count, returned, limit, names]
          properties:
            ns:
              type: string
            store_key:
              type: string
            ttl_remaining:
              type: integer
            count:
              type: integer
            returned:
              type: integer
            limit:
              type: integer
            names:
              type: array
              items:
                type: string
    DeleteStoreResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [deleted]
          properties:
            deleted:
              type: boolean
    LogsResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [lines, returned, tail]
          properties:
            lines:
              type: array
              items:
                type: string
            returned:
              type: integer
            tail:
              type: integer
    ConfigResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required:
            - backend_version
            - er_prefix
            - ttl_max_sec
            - default_limit
            - max_query_limit
            - store_preview_limit
          properties:
            backend_version:
              type: string
            er_prefix:
              type: string
            ttl_max_sec:
              type: integer
            default_limit:
              type: integer
            max_query_limit:
              type: integer
            store_preview_limit:
              type: integer
    BitmapsResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [schema, meta, groups, defaults, count, items, document]
          properties:
            schema:
              type: string
            meta:
              type: object
              additionalProperties: true
            groups:
              type: object
              additionalProperties: true
            defaults:
              type: object
              additionalProperties: true
            count:
              type: integer
            items:
              type: array
              items:
                $ref: "#/components/schemas/BitmapItem"
            document:
              $ref: "#/components/schemas/BitmapsDocument"
    BitmapsDocument:
      type: object
      required: [schema, meta, groups, labels, defaults, items, ranges]
      properties:
        schema:
          type: string
        meta:
          type: object
          additionalProperties: true
        groups:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/BitmapGroup"
        labels:
          type: object
          additionalProperties: true
        defaults:
          type: object
          additionalProperties: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/BitmapDocumentItem"
        ranges:
          type: array
          items:
            $ref: "#/components/schemas/BitmapRange"
    BitmapGroup:
      type: object
      required: [label, order]
      properties:
        label:
          type: string
        order:
          type: integer
        color:
          type: string
          nullable: true
    BitmapDocumentItem:
      type: object
      required: [bit]
      properties:
        bit:
          $ref: "#/components/schemas/Bit"
        key:
          type: string
        name:
          type: string
        group:
          type: string
        description:
          type: string
    BitmapRange:
      type: object
      required: [from, to]
      properties:
        from:
          $ref: "#/components/schemas/Bit"
        to:
          $ref: "#/components/schemas/Bit"
        group:
          type: string
        name_prefix:
          type: string
        format:
          type: string
        description:
          type: string
    BitmapItem:
      type: object
      required: [bit, key, name, group, description, source]
      properties:
        bit:
          $ref: "#/components/schemas/Bit"
        key:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        group:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        source:
          type: string
    NamespacesResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          const: true
        data:
          type: object
          required: [schema, default, namespaces]
          properties:
            schema:
              type: string
            default:
              type: string
            namespaces:
              type: array
              items:
                $ref: "#/components/schemas/NamespaceEntry"
            meta:
              type: object
              additionalProperties: true
    NamespaceEntry:
      type: object
      required: [id, label, prefix]
      properties:
        id:
          type: string
        label:
          type: string
        prefix:
          type: string